Move not user facing types to internal package
